+++ models/classes/runner/synchronisation/action/Move.php
@@ -17,25 +17,26 @@
  * Copyright (c) 2017 (original work) Open Assessment Technologies SA ;
  */
 
-namespace oat\taoQtiTest\models\runner\offline\action;
+namespace oat\taoQtiTest\models\runner\synchronisation\action;
 
-use oat\taoQtiTest\models\runner\offline\TestRunnerAction;
+use oat\taoQtiTest\models\runner\synchronisation\TestRunnerAction;
+use oat\taoQtiTest\models\runner\QtiRunnerServiceContext;
 
 /**
- * Class Skip
+ * Class Move
  *
- * Skip item into the test context.
+ * Move forward or back item into the test context.
  *
- * @package oat\taoQtiTest\models\runner\offline\action
+ * @package oat\taoQtiTest\models\runner\synchronisation\action
  */
-class Skip extends TestRunnerAction
+class Move extends TestRunnerAction
 {
     /**
-     * Process the skip action.
+     * Process the move action.
      *
      * Validate required fields.
-     * Stop timer.
-     * Wrap the skip to runner service.
+     * Stop/Start timer and save item state.
+     * Save item response and wrap the move to runner service.
      * Persist service context.
      * Start next timer.
      *
@@ -46,17 +47,25 @@ class Skip extends TestRunnerAction
         $this->validate();
 
         $ref       = ($this->getRequestParameter('ref') === false) ? null : $this->getRequestParameter('ref');
-        $itemDuration = null;
-        $consumedExtraTime = null;
-
-        $scope = $this->getRequestParameter('scope');
-        $start = ($this->getRequestParameter('start') !== false);
+        $direction = $this->getRequestParameter('direction');
+        $scope     = $this->getRequestParameter('scope');
+        $start     = ($this->getRequestParameter('start') !== false);
 
         try {
-            $serviceContext = $this->getServiceContext();
-            $this->getRunnerService()->endTimer($serviceContext, $itemDuration, $consumedExtraTime, $this->getStart());
 
-            $result = $this->getRunnerService()->skip($serviceContext, $scope, $ref);
+            /** @var QtiRunnerServiceContext $serviceContext */
+            $serviceContext = $this->getServiceContext(false);
+
+            if (!$this->getRunnerService()->isTerminated($serviceContext)) {
+                $this->endItemTimer($this->getStart());
+                $this->saveItemState();
+            }
+            $this->initServiceContext();
+
+            $this->saveItemResponses(false);
+
+            $serviceContext->getTestSession()->initItemTimer($this->getStart());
+            $result = $this->getRunnerService()->move($serviceContext, $direction, $scope, $ref);
 
             $response = [
                 'success' => $result,
@@ -66,14 +75,19 @@ class Skip extends TestRunnerAction
                 $response['testContext'] = $this->getRunnerService()->getTestContext($serviceContext);
             }
 
+            \common_Logger::d('Test session state : ' . $serviceContext->getTestSession()->getState());
+
             $this->getRunnerService()->persist($serviceContext);
 
-            if ($start == true) {
+            if ($start === true) {
+
                 // start the timer only when move starts the item session
                 // and after context build to avoid timing error
                 $this->getRunnerService()->startTimer($serviceContext, $this->getStart());
             }
+
         } catch (\Exception $e) {
+            \common_Logger::e($e->getMessage());
             $response = $this->getErrorResponse($e);
         }
 
@@ -81,12 +95,12 @@ class Skip extends TestRunnerAction
     }
 
     /**
-     * Scope parameter is required.
+     * Direction and scope are required for move action.
      *
      * @return array
      */
     protected function getRequiredFields()
     {
-        return array_merge(parent::getRequiredFields(), ['scope']);
+        return array_merge(parent::getRequiredFields(), ['direction', 'scope']);
     }
 }
