+++ models/classes/runner/synchronisation/action/StoreTraceData.php
@@ -17,16 +17,16 @@
  * Copyright (c) 2017 (original work) Open Assessment Technologies SA ;
  */
 
-namespace oat\taoQtiTest\models\runner\offline\action;
+namespace oat\taoQtiTest\models\runner\synchronisation\action;
 
 use oat\oatbox\event\EventManager;
 use oat\taoQtiTest\models\event\TraceVariableStored;
-use oat\taoQtiTest\models\runner\offline\TestRunnerAction;
+use oat\taoQtiTest\models\runner\synchronisation\TestRunnerAction;
 
 /**
  * Class StoreTraceData
  *
- * @package oat\taoQtiTest\models\runner\offline\action
+ * @package oat\taoQtiTest\models\runner\synchronisation\action
  */
 class StoreTraceData extends TestRunnerAction
 {
